Added overloaded Enqueue(Action action)
This little convenience make the simplest usage a one liner:
`UnityMainThreadDispatcher.Instance().Enqueue(() => Debug.Log ("This is executed from the main thread"));`

I've implemented this in my own project and it seems to be working just fine. I'm no Unity expert, so if there are any risks to this I may not be able to tell.